requirements.txt
python-telegram-bot==20.7
openai==1.101.0
httpx==0.27.0
